remove unused reducer code in cryptos and rename transaction to transfer to better distinguish
