+++ src/reducers/index.js
@@ -2,7 +2,7 @@ import { combineReducers } from 'redux'
 
 import accounts from './accounts'
 import cryptos from './cryptos'
-import transaction from './transaction'
+import transfer from './transfer'
 import transactions from './transactions'
 
-export default combineReducers({ accounts, transactions, transaction, cryptos })
+export default combineReducers({ accounts, transactions, transfer, cryptos })
